requirements.txt
flask
flask-cors
python-dotenv
pulp
pandas
numpy
